% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_deaths.r
\name{allocate_deaths}
\alias{allocate_deaths}
\title{Randomly allocates deaths in the stochastic CWD model}
\usage{
allocate_deaths(deaths, pop)
}
\arguments{
\item{deaths}{A vector of how many die in each age category}

\item{pop}{A matrix of age.categories = rows and number of I categories = columns}
}
\value{
A new matrix of the # of individuals in each I and age category
}
\description{
Randomly allocates deaths in the stochastic CWD model
}
