% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_det_model_fxn.r
\name{cwd_det_model}
\alias{cwd_det_model}
\title{CWD deterministic model function}
\usage{
cwd_det_model(params)
}
\arguments{
\item{params}{A list of parameters}
}
\value{
A list with 2 dataframes: counts-the # of individuals in S, I and age
categories over time. deaths--how individuals died over time.
(hunting, natural or disease)
}
\description{
Deterministic monthly age and sex structured model constant
 environmental transmission, dynamic direct transmission
}
