% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_det_model_wiw_fxn.r
\name{cwd_det_model_wiw}
\alias{cwd_det_model_wiw}
\title{CWD who-infects-who deterministic model}
\usage{
cwd_det_model_wiw(params)
}
\arguments{
\item{params}{A list with the following parameters included: 

fawn.an.sur = annual fawn survival (scaler value between 0 and 1),  

juv.an.sur = annual juvenile survival (scaler value between 0 and 1),  

ad.an.f.sur = annual adult female survival (scaler value between 0 and 1),  

ad.an.m.sur = annual adult male survival (scaler value between 0 and 1),   

fawn.repro = fawn reproduction (scaler value >= 0),  

juv.repro = juvenile reproduction (scaler value >= 0),  

ad.repro = adult reproduction (scaler value >= 0),  

hunt.mort.fawn = percentage of fawns hunted (scaler value between 0 and 1),  

hunt.mort.juv.f = percentage of juvenile females hunted (scaler value between 0 and 1),  

hunt.mort.juv.m = percentage of juvenile males hunted (scaler value between 0 and 1),  

hunt.mort.ad.f = percentage of adult females hunted (scaler value between 0 and 1),  

hunt.mort.ad.m = percentage of adult males hunted (scaler value between 0 and 1),  

ini.fawn.prev = percentage of fawns infected at the start (scaler value between 0 and 1),  

ini.juv.prev = percentage of juveniles infected at the start (scaler value between 0 and 1),  

ini.ad.f.prev = percentage of adult females infected at the start (scaler value between 0 and 1),  

ini.ad.m.prev = percentage of adult males infected at the start (scaler value between 0 and 1),   

n.age.cats = number of age categories to monitor. (scaler value greater than 3).
The final age category includes all those of that age or greater. 

p = rate of movement in the infectious categories (scaler values between 0 and 1). 
See model documentation vignette for how this relates to disease induced mortality.   

env.foi = % of the population that is infected by the environment per year (scaler value between 0 and 1) #' 
beta.ff = female to female transmission coefficient (scaler value greater than 0),  

gamma.mm = relative increase in the male-male transmission coefficient compared 
to female-female transmission (scaler value greater than 0. A value of 1 indicates equal transmission)  

gamma.mf = relative increase in the male-female transmission coefficient compared 
to female-female transmission (scaler value greater than 0. A value of 1 indicates equal transmission)  

gamma.fm = relative increase in the female-male transmission coefficient compared 
to female-female transmission (scaler value greater than 0. A value of 1 indicates equal transmission) 

theta = effect of population size on transmission (1 = frequency dependence, 0 = density dependent).  

n0 = initial population size (scaler value greater than 0)

n.years = number of years to run the model (scaler value greater than 2),  

rel.risk = relative risk of infected individuals being hunted. A value of 1 
indicates no hunter preference for infected individuals}
}
\value{
A list with 2 dataframes is returned as output: 

1. counts of the # of individuals in the susceptible and infectious 
categories by over time. 

 Columns include: 

 age (in years)

 month of simulation,

 population = number of individuals

 category: St.f = susceptible females, St.m = susceptible males, Ixt.f = 
 infectious females in the x category (1-10), Ixt.m = infectious males in the 
 x infectious category (1-10) 

 sex = female or male

disease = yes or no for susceptible or infectious

  
2. deaths--how individuals died over time (hunting, natural or disease).

 Columns include: 
 
 age in years, 
 
 month of the simulation, 
 
 population = # of individuals, 
 
 category: Ht.f = hunted females, Ht.m = hunted males, Dt.f = natural 
 mortality females, Dt.m = natural mortality males, CWDt.f = disease mortality 
 females, CWDt.m = disease mortality males.  
   
 year = year of the simulation
 
 sex
}
\description{
Deterministic monthly age and sex structured model with constant
 environmental transmission and dynamic direct transmission. 2x2 matrix of 
 transmission rates between males and females.
}
\examples{
params <- list(fawn.an.sur = 0.6, juv.an.sur = 0.8, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.1, hunt.mort.ad.m = 0.2, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04,  ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.ff = 0.06, 
gamma.mm = 2, gamma.mf = 2, gamma.fm = 1,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0)

out <- cwd_det_model_wiw(params)

plot_tots(out$counts)

}
