% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd_stoch_model_fxn.r
\name{cwd_stoch_model}
\alias{cwd_stoch_model}
\title{CWD stochastic model function}
\usage{
cwd_stoch_model(params)
}
\arguments{
\item{params}{A list of parameters}
}
\value{
A list with 2 dataframes: counts-the # of individuals in S, I and age
categories over time. deaths--how individuals died over time.
(hunting, natural or disease)
}
\description{
Stochastic monthly age and sex structured model with constant
 environmental transmission, dynamic direct transmission
}
