% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWDsims.r
\docType{package}
\name{CWDsims}
\alias{CWDsims}
\alias{CWDsims-package}
\title{CWDsims: A package for simulating chronic wasting disease models.}
\description{
Description here
}
\section{Plotting functions}{

The plotting functions ...
}

\section{model functions}{
NA
}

\section{helper functions}{
NA
}

