% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_beta_params.r
\name{est_beta_params}
\alias{est_beta_params}
\title{Beta distribution conversion function}
\usage{
est_beta_params(mu, var)
}
\arguments{
\item{mu}{mean}

\item{var}{variance}
}
\value{
A list of alpha and beta values
}
\description{
Converts mean and variance parameters to the shape and scale parameters of a
Beta distribution
}
