% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R, R/launchCWDapp.R
\name{launchApp}
\alias{launchApp}
\title{launches the shinyCWDApp}
\usage{
launchApp(example)

launchApp(example)
}
\arguments{
\item{example}{name of the application to run}

\item{example}{name of the application to run}
}
\description{
launches the shinyCWDApp

launches the shinyCWDApp
}
