% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCWDapp.R
\name{launchCWDapp}
\alias{launchCWDapp}
\title{launches the shinyCWDApp}
\usage{
launchCWDapp(example)
}
\arguments{
\item{example}{name of the application to run}
}
\description{
launches the shinyCWDApp
}
