% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_dist.r
\name{plot_age_dist}
\alias{plot_age_dist}
\title{Plot the age distribution at the end}
\usage{
plot_age_dist(dat, ...)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}
}
\value{
a plot the age distribution at the end point
}
\description{
Plot the age distribution at the end
}
