% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_buck_doe.r
\name{plot_buck_doe}
\alias{plot_buck_doe}
\title{Plot the buck:doe ratio}
\usage{
plot_buck_doe(dat, ...)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}
}
\value{
a plot the buck:doe ratio
}
\description{
Plot the buck:doe ratio
}
