% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_all.r
\name{plot_compare_all}
\alias{plot_compare_all}
\title{Stochastic Comparison plot}
\usage{
plot_compare_all(outa, outb, ...)
}
\arguments{
\item{outa}{counts as provided as output from the CWD model functions for the
first simulation}

\item{outb}{counts as provided as output from the CWD model functions for the
second simulation}
}
\value{
a density plot comparison
}
\description{
Stochastic Comparison plot
}
