% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_all_det.r
\name{plot_compare_all_det}
\alias{plot_compare_all_det}
\title{Deterministic Comparison plot}
\usage{
plot_compare_all_det(outa, outb, ...)
}
\arguments{
\item{outa}{counts as provided as output from the CWD model functions for the
first simulation}

\item{outb}{counts as provided as output from the CWD model functions for the
second simulation}
}
\value{
a bar plot comparison
}
\description{
Deterministic Comparison plot
}
