% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_hunted.r
\name{plot_compare_hunted}
\alias{plot_compare_hunted}
\title{Stochastic Comparison plot of all hunted}
\usage{
plot_compare_hunted(outa, outb, end, males.only, old.only)
}
\arguments{
\item{outa}{deaths data as provided as output from the CWD model functions for the
first simulation}

\item{outb}{deaths data as provided as output from the CWD model functions for the
second simulation}

\item{end}{TRUE/FALSE for whether to show just the last timepoint
(end = TRUE), or the cumulative number over the whole simulation.
Default = FALSE.}

\item{males.only}{TRUE/FALSE for whether to show only males
Default = FALSE.}

\item{old.only}{TRUE/FALSE for whether to show just those hunted over 3yrs.
Default = FALSE.}
}
\value{
a density plot comparison
}
\description{
Stochastic Comparison plot of all hunted
}
\examples{
params.a <- list(fawn.an.sur = 0.6, juv.an.sur = 0.8, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.2, hunt.mort.ad.m = 0.2, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04,  ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.f = 0.15,  beta.m = 0.15,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0, 
repro.var = 0.005, fawn.sur.var = 0.005, sur.var = 0.005, hunt.var = 0.005)

params.b <- list(fawn.an.sur = 0.6, juv.an.sur = 0.8, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.05, hunt.mort.ad.m = 0.6, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04,  ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.f = 0.15,  beta.m = 0.15,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0, 
repro.var = 0.005, fawn.sur.var = 0.005, sur.var = 0.005, hunt.var = 0.005)
out.a <- cwd_stoch_wrapper(params.a, nsims = 20)
out.b <- cwd_stoch_wrapper(params.b, nsims = 20)

plot_compare_hunted(out.a$deaths, out.b$deaths)

}
