% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_prev.r
\name{plot_compare_prev}
\alias{plot_compare_prev}
\title{Stochastic Comparison plot of prevalence over time}
\usage{
plot_compare_prev(outa, outb, ...)
}
\arguments{
\item{outa}{counts as provided as output from the CWD model functions for the
first simulation}

\item{outb}{counts as provided as output from the CWD model functions for the
second simulation}
}
\value{
a comparison plot of prevalence
}
\description{
Stochastic Comparison plot of prevalence over time
}
