% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_tots.r
\name{plot_compare_tots}
\alias{plot_compare_tots}
\title{Stochastic Comparison plot of totals over time}
\usage{
plot_compare_tots(outa, outb, ...)
}
\arguments{
\item{outa}{counts as provided as output from the CWD model functions for the
first simulation}

\item{outb}{counts as provided as output from the CWD model functions for the
second simulation}
}
\value{
a density plot comparison
}
\description{
Stochastic Comparison plot of totals over time
}
