% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deaths.r
\name{plot_deaths}
\alias{plot_deaths}
\title{Plot the deaths by category.}
\usage{
plot_deaths(dat, percents)
}
\arguments{
\item{dat}{deaths as provided as output from the CWD model functions}

\item{percents}{TRUE/FALSE on whether to plot the totals or the percentage
absolute totals are the default}
}
\value{
a plot deaths by category
}
\description{
Plot the deaths by category.
}
