% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fawn_doe.r
\name{plot_fawn_doe}
\alias{plot_fawn_doe}
\title{Plot the fawn to doe ratio over time}
\usage{
plot_fawn_doe(dat, ...)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}
}
\value{
a plot the fawn:doe ratio
}
\description{
Plot the fawn to doe ratio over time
}
