% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prev_age.r
\name{plot_prev_age}
\alias{plot_prev_age}
\title{Plot the prevalence over time by age}
\usage{
plot_prev_age(dat, by.sex, ...)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}

\item{by.sex}{TRUE or FALSE on whether to facet by sex}
}
\value{
a plot of the prevalence over time
}
\description{
Allows for subsetting by sex
}
