% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prev_age_end.r
\name{plot_prev_age_end}
\alias{plot_prev_age_end}
\title{Plot the prevalence by age at the end of the simulation}
\usage{
plot_prev_age_end(dat, ...)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}
}
\value{
a plot of the prevalence by age
}
\description{
Plot the prevalence by age at the end of the simulation
}
