% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prev_time.r
\name{plot_prev_time}
\alias{plot_prev_time}
\title{Plot the prevalence over time}
\usage{
plot_prev_time(dat, ...)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}
}
\value{
a plot of the prevalence over time
}
\description{
Plot the prevalence over time
}
