% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_age_dist.r
\name{plot_stoch_age_dist}
\alias{plot_stoch_age_dist}
\title{Age distribution plot}
\usage{
plot_stoch_age_dist(dat, ...)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}
}
\value{
a plot of the percent of the population in each age class
}
\description{
Age distribution plot
}
