% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_buck_doe.r
\name{plot_stoch_buck_doe}
\alias{plot_stoch_buck_doe}
\title{Buck:doe stochastic plot}
\usage{
plot_stoch_buck_doe(dat, all.lines, error.bars, ...)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{all.lines}{TRUE/FALSE for whether to plot a line for every simulation}

\item{error.bars}{error bars = vector of high and low percentiles
(2 values only). If missing, no error bars are shown.}
}
\value{
a plot of the ratio of adult males to adult females over time
}
\description{
Buck:doe stochastic plot
}
