% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_deaths.r
\name{plot_stoch_deaths}
\alias{plot_stoch_deaths}
\title{Death types stochastic plot}
\usage{
plot_stoch_deaths(dat, error.bars)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{error.bars}{error bars = vector of high and low percentiles
(2 values only). If missing, no error bars are shown.}
}
\value{
a plot of death types over time.
}
\description{
Death types stochastic plot
}
