% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_disease.r
\name{plot_stoch_disease}
\alias{plot_stoch_disease}
\title{Plot stochastic totals of positives and negatives over time}
\usage{
plot_stoch_disease(dat, error.bars)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{error.bars}{error bars = vector of high and low percentiles
(2 values only). If missing, no error bars are shown.}
}
\value{
a multiple line plot of the simulation over time
}
\description{
Plot stochastic totals of positives and negatives over time
}
