% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_fawn_doe.r
\name{plot_stoch_fawn_doe}
\alias{plot_stoch_fawn_doe}
\title{Fawn:doe stochastic plot}
\usage{
plot_stoch_fawn_doe(dat, all.lines, error.bars, ...)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{all.lines}{TRUE/FALSE for whether to plot a line for every simulation}

\item{error.bars}{2 value vector for the hi and lo percentiles on the error
bars. If missing, no error bars will be shown.}
}
\value{
a plot of the percent of the population in each age class
}
\description{
Fawn:doe stochastic plot
}
