% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_perc_deaths.r
\name{plot_stoch_perc_deaths}
\alias{plot_stoch_perc_deaths}
\title{Percentage death types stochastic plot}
\usage{
plot_stoch_perc_deaths(dat, error.bars)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{error.bars}{2 value vector for the hi and lo percentiles on the error
bars. If missing, no error bars are shown.}
}
\value{
a plot of % death types over time.
}
\description{
Percentage death types stochastic plot
}
