% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_prev.R
\name{plot_stoch_prev}
\alias{plot_stoch_prev}
\title{Plot stochastic prevalence over time}
\usage{
plot_stoch_prev(dat, all.lines, error.bars, ...)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{all.lines}{TRUE/FALSE for whether to plot a line for every simulation.
Default = TRUE.}

\item{error.bars}{2 value vector for the hi and lo percentiles on the error
bars. If missing, no error bars are shown}
}
\value{
a multiple line plot of the simulation over time
}
\description{
Plot stochastic prevalence over time
}
