% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_prev_age_end.r
\name{plot_stoch_prev_age_end}
\alias{plot_stoch_prev_age_end}
\title{Prevalence versus age plot}
\usage{
plot_stoch_prev_age_end(dat, error.bars, ...)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{error.bars}{vector with 2 values for the low and high percentiles. If
missing, then no error bars will be shown.}
}
\value{
a plot of prevalence versus age at the last timepoint of the
simulation
}
\description{
Prevalence versus age plot
}
