% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_tots.r
\name{plot_stoch_tots}
\alias{plot_stoch_tots}
\title{Plot stochastic totals over time}
\usage{
plot_stoch_tots(dat, all.lines, error.bars, by.sexage)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{all.lines}{TRUE/FALSE for whether to plot a line for every simulation}

\item{error.bars}{error bars = vector of high and low percentiles (2 values only).
If missing, no error bars will be shown.}

\item{by.sexage}{TRUE/FALSE for whether to facet by sex and age.}
}
\value{
a multiple line plot of the simulation over time
}
\description{
Plot stochastic totals over time
}
\examples{
params <- list(fawn.an.sur = 0.6, juv.an.sur = 0.8, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.1, hunt.mort.ad.m = 0.4, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04,  ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.f = 0.15,  beta.m = 0.15,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0, 
repro.var = 0.005, fawn.sur.var = 0.005, sur.var = 0.005, hunt.var = 0.005)

out <- cwd_stoch_wrapper(params, nsims = 20)
plot_stoch_tots(out$counts, all.lines = TRUE, by.sexage = TRUE, 
error.bars = c(0.05, 0.95))

}
