% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tots.r
\name{plot_tots}
\alias{plot_tots}
\title{Plot the total of S and I over time}
\usage{
plot_tots(dat, ...)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}
}
\value{
a plot of the population totals split by age.
}
\description{
Plot the total of S and I over time
}
