% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ttd.r
\name{plot_ttd}
\alias{plot_ttd}
\title{Plot Time to disease-induced death}
\usage{
plot_ttd(p)
}
\arguments{
\item{p}{proportion or probability of progressing through the infections
categories. Must be between 0 and 1.}
}
\value{
a density plot of time to death.
}
\description{
Creates a plot of the time (in years) to disease-induced death based on the
proportion or probability of movement through the 10 infectious
sub-categories. Uses 1000 draws from a Gamma distribution.
}
\examples{
plot_ttd(p = 0.043)

}
