% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vitals.r
\name{plot_vitals}
\alias{plot_vitals}
\title{Vital rate plot}
\usage{
plot_vitals(params)
}
\arguments{
\item{params}{list of the parameters provided to the CWD model}
}
\value{
a plot of the vital rate distributions
}
\description{
Creates a plot of the survival and reproduction distributions that are
defined by the CWD model parameters
}
