% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_buck_doe.r
\name{plot_buck_doe}
\alias{plot_buck_doe}
\title{Plot the buck:doe ratio}
\usage{
plot_buck_doe(dat)
}
\arguments{
\item{dat}{counts provided as output from the CWD model functions}
}
\value{
a plot the buck:doe ratio
}
\description{
Plot the buck:doe ratio
}
\examples{
params <- list(fawn.an.sur = 0.6, juv.an.sur = 0.8, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.2, hunt.mort.ad.m = 0.4, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04,  ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.f = 0.15,  beta.m = 0.15,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0)

out <- cwd_det_model(params)
plot_buck_doe(out$counts)

}
