% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_all_det.r
\name{plot_compare_all_det}
\alias{plot_compare_all_det}
\title{Deterministic Comparison plot}
\usage{
plot_compare_all_det(outa, outb)
}
\arguments{
\item{outa}{counts as provided as output from the CWD model functions for the
first simulation}

\item{outb}{counts as provided as output from the CWD model functions for the
second simulation}
}
\value{
a bar plot comparison
}
\description{
Deterministic Comparison plot
}
\examples{
params.a <- list(fawn.an.sur = 0.6, juv.an.sur = 0.8, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.2, hunt.mort.ad.m = 0.2, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04,  ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.f = 0.15,  beta.m = 0.15,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0)

params.b <- list(fawn.an.sur = 0.6, juv.an.sur = 0.8, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.1, hunt.mort.ad.m = 0.5, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04, ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.f = 0.15,  beta.m = 0.15,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0)

out.a <- cwd_det_model(params.a)
out.b <- cwd_det_model(params.b)

plot_compare_all_det(out.a, out.b)

}
