% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_hunted.r
\name{plot_compare_hunted}
\alias{plot_compare_hunted}
\title{Stochastic Comparison plot of all hunted}
\usage{
plot_compare_hunted(outa, outb, end, males.only, old.only, ...)
}
\arguments{
\item{outa}{counts as provided as output from the CWD model functions for the
first simulation}

\item{outb}{counts as provided as output from the CWD model functions for the
second simulation}

\item{end}{TRUE/FALSE for whether to show just the last timepoint
(end = TRUE), or the cumulative number over the whole simulation.
Default = FALSE.}

\item{males.only}{TRUE/FALSE for whether to show only males
Default = FALSE.}

\item{old.only}{TRUE/FALSE for whether to show just those hunted over 3yrs.
Default = FALSE.}
}
\value{
a density plot comparison
}
\description{
Stochastic Comparison plot of all hunted
}
