% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_prev_age.r
\name{plot_stoch_prev_age}
\alias{plot_stoch_prev_age}
\title{Plot stochastic prevalence by age over time}
\usage{
plot_stoch_prev_age(dat, by.sex, ...)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{by.sex}{TRUE/FALSE on whether to facet by sex. Default = FALSE}
}
\value{
a multiple line plot of the simulation over time
}
\description{
Plot stochastic prevalence by age over time
}
