% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stoch_tots.r
\name{plot_stoch_tots}
\alias{plot_stoch_tots}
\title{Plot stochastic totals over time}
\usage{
plot_stoch_tots(dat, all.lines, error.bars, by.sexage, ...)
}
\arguments{
\item{dat}{counts as provided as output from the CWD model}

\item{all.lines}{TRUE/FALSE for whether to plot a line for every simulation}

\item{error.bars}{error bars = vector of high and low percentiles (2 values only).
If missing, no error bars will be shown.}

\item{by.sexage}{TRUE/FALSE for whether to facet by sex and age.}
}
\value{
a multiple line plot of the simulation over time
}
\description{
Plot stochastic totals over time
}
