% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vitals.r
\name{plot_vitals}
\alias{plot_vitals}
\title{Vital rate plot}
\usage{
plot_vitals(params)
}
\arguments{
\item{params}{list of the parameters provided to the CWD model}
}
\value{
a plot of the vital rate distributions
}
\description{
Creates a plot of the survival and reproduction distributions that are
defined by the CWD model parameters
}
\examples{

params <- list(fawn.an.sur = 0.4, juv.an.sur = 0.6, ad.an.f.sur = 0.95, 
ad.an.m.sur = 0.9, fawn.repro = 0, juv.repro = 0.6, ad.repro = 1, 
hunt.mort.fawn = 0.01, hunt.mort.juv.f = 0.1, hunt.mort.juv.m = 0.1,
hunt.mort.ad.f = 0.2, hunt.mort.ad.m = 0.4, ini.fawn.prev = 0.02,
ini.juv.prev = 0.03, ini.ad.f.prev = 0.04,  ini.ad.m.prev = 0.04,
n.age.cats = 12,  p = 0.43, env.foi = 0,  beta.f = 0.15,  beta.m = 0.15,
theta = 1, n0 = 2000, n.years = 10, rel.risk = 1.0,  
repro.var = 0.005, fawn.sur.var = 0.005, sur.var = 0.005, hunt.var = 0.005)

plot_vitals(params = params)
}
